Build script, heap stats, and hello world (yo)
Create a C++ object, see that stuff gets allocated in the heap, and then delete it, and see that everything gets freed.  This demonstrates that we are using dynamic memory allocation and don’t have leaks, and that we have the stats to reason about how we are using the Emscripten heap to store instances of PbSolver.