Use a “bridge” w/ MiniSAT to solve incrementally
We insert SolverBridge as a bridge interface between MiniSat+ and MiniSat.  This aids debugging a lot, and it also lets use ensure that we are able to add PB constraints which are “guarded” behind a variable.  We can then use this variable as an MiniSat assumption in order to temporarily enforce a PB constraint.